feat(vs2022_third): C4244 fix double→long conversions in RecipeMainSrView.cpp — 1 file
Fix 26 C4244 warnings (double→long narrowing conversions) by adding
static_cast<long>() at assignment/argument sites.

Lines fixed: 3954, 3962, 4362, 4363, 4368, 4760, 4763, 4768, 4771,
4868, 4906, 4907, 4910, 4911, 4914, 4915, 4918, 4919, 4925, 4926,
4929, 4930, 4933, 4934, 4937, 4938.

No behavior changes, only explicit casts to match target types.

+++ SRC/NanoSpec/RecipeMainSrView.cpp
@@ -3951,7 +3951,7 @@ void CRecipeMainSrView::OnPaint()
 
 	if((dStageMaxSizeX / dStageMaxSizeY) > ((StageRectRight - StageRectLeft) / (StageRectBottom - StageRectTop))) {
 		lDrawStageSizeX =(StageRectRight - StageRectLeft);
-		lDrawStageSizeY = ((StageRectRight - StageRectLeft) * dStageMaxSizeY / dStageMaxSizeX);
+		lDrawStageSizeY = static_cast<long>((StageRectRight - StageRectLeft) * dStageMaxSizeY / dStageMaxSizeX);
 
 		m_dDrawStartStagePosX = StageRectLeft;
 		m_dDrawStartStagePosY = StageRectTop + ((StageRectBottom - StageRectTop - lDrawStageSizeY) / 2);
@@ -3959,7 +3959,7 @@ void CRecipeMainSrView::OnPaint()
 		m_dDrawEndStagePosY = m_dDrawStartStagePosY + lDrawStageSizeY;
 	}
 	else{
-		lDrawStageSizeX = ((StageRectBottom - StageRectTop) * dStageMaxSizeX / dStageMaxSizeY);
+		lDrawStageSizeX = static_cast<long>((StageRectBottom - StageRectTop) * dStageMaxSizeX / dStageMaxSizeY);
 		lDrawStageSizeY = (StageRectBottom - StageRectTop);
 
 		m_dDrawStartStagePosX = StageRectLeft + ((StageRectRight - StageRectLeft - lDrawStageSizeX) / 2);
@@ -4359,13 +4359,13 @@ void CRecipeMainSrView::DrawSample(CDC* pDC, int iOrg, int DirX, int DirY)
 				double dDeltaGraphSize_x = dSampleSizeX * (dGraphSize_y / dSampleSizeY);
 				//スペースのセンターあわせ用座標計算
 				double dGraphCenterOffset_x = (dGraphSize_x - dDeltaGraphSize_x)/2;
-				placement.rcNormalPosition.left = placement.rcNormalPosition.left + dGraphCenterOffset_x;
-				placement.rcNormalPosition.right = placement.rcNormalPosition.left + dDeltaGraphSize_x;
+				placement.rcNormalPosition.left = placement.rcNormalPosition.left + static_cast<LONG>(dGraphCenterOffset_x);
+				placement.rcNormalPosition.right = placement.rcNormalPosition.left + static_cast<LONG>(dDeltaGraphSize_x);
 			}
 			else{
 				//Yサイズ設定
 				double dDeltaGraphSize_y = dSampleSizeY * (dGraphSize_x / dSampleSizeX);
-				placement.rcNormalPosition.bottom = placement.rcNormalPosition.top + dDeltaGraphSize_y;
+				placement.rcNormalPosition.bottom = placement.rcNormalPosition.top + static_cast<LONG>(dDeltaGraphSize_y);
 			}
 
 			//Y方向もセンターに合わせる
@@ -4757,18 +4757,18 @@ void CRecipeMainSrView::SamplePointGraph_DataSet()
 	switch ( m_StageConfig.RotateXyView ) {
 	case 0:
 	default:
-		m_SamplePointGraph.SetOriginPointData(dOriginPointDataX, dOriginPointDataY);
+		m_SamplePointGraph.SetOriginPointData(static_cast<long>(dOriginPointDataX), static_cast<long>(dOriginPointDataY));
 		break;
 	case 90:
-		m_SamplePointGraph.SetOriginPointData(dOriginPointDataY, dOriginPointDataX);
+		m_SamplePointGraph.SetOriginPointData(static_cast<long>(dOriginPointDataY), static_cast<long>(dOriginPointDataX));
 		dOffSetValueX = dOriginPointDataY;
 		dOffSetValueY = dOriginPointDataX;
 		break;
 	case 180:
-		m_SamplePointGraph.SetOriginPointData(dOriginPointDataX, dOriginPointDataY);
+		m_SamplePointGraph.SetOriginPointData(static_cast<long>(dOriginPointDataX), static_cast<long>(dOriginPointDataY));
 		break;
 	case 270:
-		m_SamplePointGraph.SetOriginPointData(dOriginPointDataY, dOriginPointDataX);
+		m_SamplePointGraph.SetOriginPointData(static_cast<long>(dOriginPointDataY), static_cast<long>(dOriginPointDataX));
 		dOffSetValueX = dOriginPointDataY;
 		dOffSetValueY = dOriginPointDataX;
 		break;
@@ -4865,7 +4865,7 @@ void CRecipeMainSrView::SamplePointGraph_DataSet()
 			Scan_PointIncOffsetY = dOffSetValueY + m_ScanPoint[iPoint].lY;
 			break;
 		}
-		m_SamplePointGraph.AddPoint(Scan_PointIncOffsetX, Scan_PointIncOffsetY);
+		m_SamplePointGraph.AddPoint(static_cast<long>(Scan_PointIncOffsetX), static_cast<long>(Scan_PointIncOffsetY));
 	}
 #endif
 // 2009.10.28 bagus 2点間 --}--
